requirements.txt
python-kasa>=0.5.1
python-rtmidi>=1.5.0 